Update SVGSprite.cpp
I have downloaded and used your library; it works very well on Windows. However, when building for Android and iOS, there is an error where the SVG file cannot be opened when using the function: nsvgParseFromFile. I switched to using the Cocos2d-x library code to open the file, and it works well on Android and iOS. Additionally, it is not necessary to include the glfw3 library to load SVG files on Android and iOS. Thank you very much, because with your sample code, I was able to load SVG files on Android and iOS